+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# Version 2.6.2 (2017-10-19)
+
+* [new] Add basic prompt mode (-DbasicPrompt) on generate goal, for the case where ConsoleUI doesn't work properly.
+* [chg] If an archetype named "web" exists, set it as the default project type.
+
 # Version 2.6.2 (2017-08-10)
 
 * [chg] Add a fallback to hard-coded archetype list for project generation.